Move threading from pyanaconda to pyanaconda.core
