+++ pyanaconda/startup_utils.py
@@ -48,7 +48,7 @@ from pyanaconda.modules.common.constants.objects import STORAGE_CHECKER
 from pyanaconda.modules.common.constants.services import TIMEZONE, LOCALIZATION, SERVICES, \
     SECURITY, STORAGE
 from pyanaconda.modules.common.util import is_module_available
-from pyanaconda.threading import AnacondaThread, thread_manager
+from pyanaconda.core.threading import AnacondaThread, thread_manager
 
 stdout_log = get_stdout_logger()
 log = get_module_logger(__name__)
